+++ packages/instantsearch-react/src/impl/Hits.test.js
@@ -2,28 +2,14 @@
 
 import React from 'react';
 import {mount} from 'enzyme';
+import renderer from 'react-test-renderer';
 
 import Hits from './Hits';
-import DefaultHitComponent from './DefaultHitComponent';
 jest.unmock('./Hits');
-jest.unmock('./DefaultHitComponent');
+jest.unmock('../themeable');
+jest.unmock('../propTypes');
 
 describe('Hits', () => {
-  // @FIX: Triggers a warning https://github.com/facebook/react/issues/7240
-  // Remove this comment once the issue is closed
-  it('renders a default component when no itemComponent prop is provided', () => {
-    const hits = [{objectID: 0}, {objectID: 1}, {objectID: 2}];
-    const wrapper = mount(
-      <Hits
-        hits={hits}
-      />
-    );
-    expect(wrapper.find(DefaultHitComponent).length).toBe(3);
-    expect(wrapper.find(DefaultHitComponent).everyWhere((c, i) =>
-      c.props().hit === hits[i]
-    )).toBe(true);
-  });
-
   it('accepts a itemComponent prop', () => {
     const hits = [{objectID: 0}, {objectID: 1}, {objectID: 2}];
     const Hit = () => null;
@@ -37,5 +23,19 @@ describe('Hits', () => {
     expect(wrapper.find(Hit).everyWhere((c, i) =>
       c.props().hit === hits[i]
     )).toBe(true);
+  });
+
+  it('applies theme', () => {
+    const tree = renderer.create(
+      <HitsPerPage
+        values={[111, 333, 666]}
+        theme={{
+          label: 'HITS_LABEL',
+          value: v => `HITS_VALUE_${v}`,
+        }}
+        hitsPerPage={111}
+      />
+    ).toJSON();
+    expect(tree).toMatchSnapshot();
   });
 });
